CSE 20
HelloWorldGUI.java - Classic Hello World exercise with a twist, implemented the Swing class to display GUI for user.
AddNumbersGUI.java - Addition exercise implemented Swing class to display GUI for user, takes two arguments and adds them together.
Sort.java - Reads the input strings, one by one, and places them in the array in Alphabetical order, using compareTo method.
Encrypt.java  - Uses for loops and arrays to read a string and write it out encrypted with a simple substitution cipher. Because of the way it works, the same program can also decrypt the encrypted output. 

Horoscope.java - Project 1: A program that issues a different horoscope for each sign of the astrological zodiac, based upon the user’s birth date. 

Cards.java - Project 2: A program that sets up a deck of cards, shuffles the deck and deals out hands, the number and size of which are specified by the user as command-line arguments.